+++ modules/mystic-clicker/entry.cpp
@@ -57,7 +57,7 @@ extern "C" __declspec(dllexport) AddonDefinition_t* GetAddonDef()
     AddonDef.Name = "Mystic Clicker";
     AddonDef.Version.Major = 3;
     AddonDef.Version.Minor = 6;
-    AddonDef.Version.Build = 14;
+    AddonDef.Version.Build = 15;
     AddonDef.Version.Revision = 0;
     AddonDef.Author = "OgMorrow2090";
     AddonDef.Description = "One-click hotkeys for inventory, vendors, trading post, Mystic Forge, and more.";
@@ -188,6 +188,7 @@ void AddonLoad(AddonAPI_t* aApi)
     APIDefs->InputBinds_RegisterWithString(GUILD_HALL, ProcessKeybind, "(null)");
     APIDefs->InputBinds_RegisterWithString(GUILD_HALL_COMBO, ProcessKeybind, "CTRL+SHIFT+F5");
     APIDefs->InputBinds_RegisterWithString(GRACEFUL_QUIT, ProcessKeybind, "ALT+SHIFT+Q");
+    APIDefs->InputBinds_RegisterWithString(PATHING_TOGGLE_ALL, ProcessKeybind, "CTRL+F3");
     APIDefs->InputBinds_RegisterWithString(CAPTURE_CRAFT, ProcessKeybind, "(null)");
     APIDefs->InputBinds_RegisterWithString(CAPTURE_CRAFT_ALL, ProcessKeybind, "(null)");
     APIDefs->InputBinds_RegisterWithString(CAPTURE_WIZARD_VAULT, ProcessKeybind, "(null)");
@@ -308,6 +309,7 @@ void AddonUnload()
     APIDefs->InputBinds_Deregister(GUILD_HALL);
     APIDefs->InputBinds_Deregister(GUILD_HALL_COMBO);
     APIDefs->InputBinds_Deregister(GRACEFUL_QUIT);
+    APIDefs->InputBinds_Deregister(PATHING_TOGGLE_ALL);
     APIDefs->InputBinds_Deregister(CAPTURE_CRAFT);
     APIDefs->InputBinds_Deregister(CAPTURE_CRAFT_ALL);
     APIDefs->InputBinds_Deregister(CAPTURE_WIZARD_VAULT);